Resolves timols/java-gitlab-api/issues/33
There is a bug in the REST API for Gitlab when retrieving commits with
pagination (https://gitlab.com/gitlab-org/gitlab-ce/issues/759). This
commit is a workaround for that - when we are retrieving commits, we
assume that the second page is numbered 1 instead of 2.

To make this not cause problems if/when the Gitlab CE project fixes the
REST API, equals() is overridden in GitlabCommit to prevent duplicate
entries.