Support OAuth access_token
For OAuth the uri query parameter should be named access_token instead of
private_token. To support both a enum TokenType is introduced and should
be used to select which token type to be used for the API access.

For backwards compatibility TokenType#PRIVATE_TOKEN will be used by
default.

Signed-off-by: Andreas Häber <andreas.haber@intele.com>