Fix getting commits for a merge request
This method used to use the merge request's source project (e.g. a fork)
to get the commits. However this appears to be unnecessary and wrong with
current GitLab versions at least 14.x, 15.x).

In particular, this would fail hard when trying to get the commits after
the source project was deleted.

The commits of a merge request belong to the project of the merge request
and not to any other project where they may have been developed in.